requirements.txt
flask
sentence-transformers
scikit-learn
google-generative-ai
requests
python-bidi
